WallaBot/requirements.txt
pandas==2.2.3
plyer==2.1.0
python-dotenv==1.0.1
selenium==4.26.1
